move `readRanges` to `lib/range`
