+++ lib/methods.mjs
@@ -12,13 +12,7 @@ import {
   READDIR_OPTS,
 } from './constants.mjs';
 import {
-  responseConstructor,
-} from './generic.mjs';
-import {
-  validatedBody,
-} from './sri.mjs';
-import {
-  readByteRanges,
+  readRanges,
   writeByteRange,
 } from './range.mjs';
 
@@ -32,34 +26,7 @@ const GET = async (url, { headers, integrity, signal }) => {
   if (conditionalResponse)
     return conditionalResponse;
 
-  const { unit, ranges } = getRanges(headers);
-
-  switch (unit) {
-  case 'bytes':
-    return readByteRanges(url, ranges, stats.size)
-      .then($ => validatedBody(integrity, $))
-      .then(
-        async body => [
-          body,
-          await statsAsOptions(
-            Object.assign(stats, { size: BigInt(body.byteLength) }),
-            { status: 206 }
-          ),
-        ]
-      )
-      .then(responseConstructor)
-      .catch($ => $ instanceof RangeError
-        ? genericResponse(416)
-        : Promise.reject($)
-      );
-  case 'none':
-    return Promise.all([
-      fs.readFile(url, { signal }).then($ => validatedBody(integrity, [$])),
-      statsAsOptions(stats),
-    ]).then(responseConstructor);
-  default:
-    return genericResponse(416);
-  }
+  return readRanges(url, { headers, integrity, signal }, stats);
 };
 
 const HEAD = async url =>
