+++ lib/range.mjs
@@ -1,4 +1,15 @@
-import { open } from 'node:fs/promises';
+import { open, readFile } from 'node:fs/promises';
+import {
+  responseConstructor,
+} from './generic.mjs';
+import {
+  validatedBody,
+} from './sri.mjs';
+import {
+  genericResponse,
+  getRanges,
+  statsAsOptions,
+} from './http.mjs';
 
 const rangeToFSOpts = (range, size = Infinity) => {
   size = Number(size);
@@ -37,6 +48,37 @@ const readByteRanges = async (url, ranges, size, fd) => {
   }
 };
 
+const readRanges = async (url, { headers, integrity, signal }, stats) => {
+  const { unit, ranges } = getRanges(headers);
+
+  switch (unit) {
+  case 'bytes':
+    return readByteRanges(url, ranges, stats.size)
+      .then($ => validatedBody(integrity, $))
+      .then(
+        async body => [
+          body,
+          await statsAsOptions(
+            {...stats, size: BigInt(body.byteLength) },
+            { status: 206 }
+          ),
+        ]
+      )
+      .then(responseConstructor)
+      .catch($ => $ instanceof RangeError
+        ? genericResponse(416)
+        : Promise.reject($)
+      );
+  case 'none':
+    return Promise.all([
+      readFile(url, { signal }).then($ => validatedBody(integrity, [$])),
+      statsAsOptions(stats),
+    ]).then(responseConstructor);
+  default:
+    return genericResponse(416);
+  }
+};
+
 const writeByteRange = async (url, range, body, flag) => {
   let { length, position } = rangeToFSOpts(range);
 
@@ -61,5 +103,6 @@ const writeByteRange = async (url, range, body, flag) => {
 
 export {
   readByteRanges,
+  readRanges,
   writeByteRange,
 };
